+++ RoughSets/Infovision.Datamining.RoughSet/IReductMeasure.cs
@@ -5,6 +5,6 @@ namespace Infovision.Datamining.Roughset
     public interface IReductMeasure : IFactoryProduct
     {
         SortDirection SortDirection { get; }
-        double Calc(IReduct reduct);
+        decimal Calc(IReduct reduct);
     }
 }
