+++ RoughSets/Infovision.Datamining.RoughSet/IReductStoreMeasure.cs
@@ -5,6 +5,6 @@ namespace Infovision.Datamining.Roughset
     public interface IReductStoreMeasure : IFactoryProduct
     {
         SortDirection SortDirection { get; }
-        double Calc(IReductStore reductStore);
+        decimal Calc(IReductStore reductStore);
     }
 }
